% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{distinct_colours}
\alias{distinct_colours}
\title{Generate n distinct colours}
\usage{
distinct_colours(n)
}
\arguments{
\item{n}{Number of colours to generate}
}
\value{
A vector of colour names.
}
\description{
Generate n distinct colours
}
