% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate_area.R
\name{integrate_area}
\alias{integrate_area}
\title{Integrate area}
\usage{
integrate_area(df, x, y, skip.na = TRUE)
}
\arguments{
\item{df}{Data frame containing \code{x} and \code{y} of a plot.}

\item{x}{Column name for the x axis.}

\item{y}{Column name for the y axis.}

\item{skip.na}{Logical. How to treat \code{NA}s in \code{y}. If \code{TRUE} the trapezium of
the area adjacent to the missing value will be calculated. If \code{FALSE} the
two trapeziums before and after the missing value will be \code{NA}.}
}
\value{
A value of the area.
}
\description{
Integrate the area under a xy plot. Credit to \href{https://stackoverflow.com/a/4955118/4227151}{StackOverflow}
The method used here is to sum the area of mutiple trapeziums under a line.
}
