% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_leaflet.R
\name{plot_leaflet}
\alias{plot_leaflet}
\title{Convenience function to quickly plot geometries on a leaflet map.}
\usage{
plot_leaflet(
  sp,
  popup = NULL,
  force = TRUE,
  colour = "#03F",
  col_group = NULL,
  n = 7,
  palette = "viridis",
  legend_pos = "topright",
  legend_title = NULL,
  opacity = 0.5,
  fill_opacity = 0.2,
  width = NULL,
  height = NULL,
  padding = 0,
  ...
)
}
\arguments{
\item{sp}{Spatial object to be plotted.}

\item{popup}{Vector of variables to be used as a popup on the map.}

\item{force}{Should the projection be forced to WGS84? Default is \code{TRUE}.}

\item{colour}{Colour of geometry.}

\item{col_group}{If the colour mapped on a numeric vector, how should the colours be grouped.
Default is \code{NULL}. Can be "bin" or "quantile".}

\item{n}{If the \verb{col_group`` is not }NULL``, how many different of colours should be plotted.}

\item{palette}{Colour pallette options passed to \code{\link[leaflet:colorNumeric]{leaflet::colorNumeric()}}.}

\item{legend_pos}{Legend position. Accepts \code{c("topright", "bottomright", "bottomleft","topleft")}.}

\item{legend_title}{Legend title.}

\item{opacity}{Opacity of the edge of the geometry.}

\item{fill_opacity}{Internal opacity of the geometry.}

\item{width}{Width of the map}

\item{height}{Height of the map}

\item{padding}{Padding of the map}

\item{...}{Other arguments passed to \code{\link[leaflet:addControl]{leaflet::addControl()}}}
}
\value{
Invisible, a leaflet map.
}
\description{
Convenience function to quickly plot geometries on a leaflet map.
}
\author{
Stuart K. Grange, Hao Wu
}
