% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_polar.R
\name{plot_polar}
\alias{plot_polar}
\title{Polar scatter plot}
\usage{
plot_polar(data, pol, ...)
}
\arguments{
\item{data}{A data frame containing wind speed and wind direction data with names
\code{ws} and \code{wd}, respectively.}

\item{pol}{The name of the pollutant to plot.}

\item{...}{Additional arguements passed to \code{\link[plotly:plot_ly]{plotly::plot_ly()}}}
}
\value{
A plotly object
}
\description{
Polar scatter plot
}
\examples{
\dontrun{
plot_polar(data, "NO2")
}


}
