% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot with plotly}
\usage{
plot_scatter(
  data,
  x,
  y,
  group = NULL,
  text = NULL,
  hoverinfo = "x+y+text",
  linear = TRUE,
  colors = "Set1",
  one2one_line = FALSE,
  eq_x = 0.6,
  eq_y = 1,
  ...
)
}
\arguments{
\item{data}{Data frame.}

\item{x}{Name of x variable.}

\item{y}{Name of y variable.}

\item{group}{Optional name for grouping variable.}

\item{text}{Optional name for text variable. This will show in the hover text box.}

\item{hoverinfo}{If text variable is specified, use "x+y+text" to show text.
For more info, check \code{\link[plotly:schema]{plotly::schema()}}}

\item{linear}{Whether a linear best fit line is drawn.}

\item{colors}{RColorBrewer palette name.}

\item{one2one_line}{Should the 1:1 line be drawn.}

\item{eq_x}{A number between 0 to 1 to specify the location of equation relative to the x axis.}

\item{eq_y}{A number between 0 to 1 to specify the location of equation relative to the y axis.}

\item{...}{Other arguments passed to \code{\link[plotly:plot_ly]{plotly::plot_ly()}}}
}
\value{
A plotly object.
}
\description{
Scatter plot with plotly
}
