% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_helpers.R
\name{prep_data}
\alias{prep_data}
\title{Prepare data for scatter plot}
\usage{
prep_data(data, x, y, group, text, na.rm = TRUE)
}
\arguments{
\item{data}{data frame.}

\item{x}{x variable}

\item{y}{y variable}

\item{group}{Optional grouping vairiable}

\item{text}{Optional text variable to show in hover info}

\item{na.rm}{remove missing x/y observations.}
}
\description{
Pad out data if group and/or text are missing
}
