% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_select.R
\name{var_select}
\alias{var_select}
\title{Select the 'best' variables to include in a linear model}
\usage{
var_select(data, y, stopat = Inf)
}
\arguments{
\item{data}{A data frame containing the predictor variables (X) and the dependent variable (Y).}

\item{y}{The name of the dependent variable.}

\item{stopat}{The total number of variables to include in the model.
Default is \code{Inf} which means using all the variables in the data frame.}
}
\value{
A nested tibble with the lm models and stats.
}
\description{
This is an implementation of a simple algorithm that selects the 'best' variables
as the predictors in a linear regression model. The selection is based on
choosing the variables that give the highest adjusted R-squared value.
}
