% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_shift.R
\name{var_shift}
\alias{var_shift}
\title{Shift variables forward or backward by some observations}
\usage{
var_shift(df, vars, n, fn)
}
\arguments{
\item{df}{Data frame containing variables to shift.}

\item{vars}{Variable(s) to shift.}

\item{n}{Number of observations to shift.}

\item{fn}{The shifting function. Usually this is \code{\link[dplyr:lead-lag]{dplyr::lead()}} or \code{\link[dplyr:lead-lag]{dplyr::lag()}}.
This could also be a custom function that takes a vector and a number \code{n}.}
}
\value{
Data frame with append shifted measurements.
}
\description{
Sometimes we want to shift some measurements backward or forward.
\code{\link[dplyr:lead-lag]{dplyr::lead()}} and \code{\link[dplyr:lead-lag]{dplyr::lag()}} do this but will only operate on a single
variable. \code{\link[=var_shift]{var_shift()}} extends this functionality and allows shifting multiple
variables by a range of steps.
}
\examples{
\dontrun{
shift(mtcars, c("vs", "am"), lead, 1:2)
}
}
