% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scatter.R
\name{plot_scatter}
\alias{plot_scatter}
\title{Scatter plot iwth plotly}
\usage{
plot_scatter(
  data,
  x,
  y,
  group = NULL,
  text = NULL,
  hoverinfo = "x+y",
  linear = TRUE
)
}
\arguments{
\item{data}{Data frame.}

\item{x}{Name of x variable.}

\item{y}{Name of y variable.}

\item{group}{Optional name for grouping variable.}

\item{text}{Optional name for text variable. This will show in the hover text box.}

\item{hoverinfo}{If text variable is specified, use "x+y+text" to show text.
For more info, check \code{\link[plotly:schema]{plotly::schema()}}}

\item{linear}{Whether a linear best fit line is drawn.}
}
\value{
A plotly object.
}
\description{
Scatter plot iwth plotly
}
